% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneR.R
\name{plot.OneR}
\alias{plot.OneR}
\title{Plot Diagnostics for an OneR object}
\usage{
\method{plot}{OneR}(x, ...)
}
\arguments{
\item{x}{object of class \code{"OneR"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plots a mosaic plot for the feature attribute and the target of the OneR model.
}
\details{
If more than 20 levels are present for either the feature attribute or the target the function stops with an error.
}
\examples{
model <- OneR(iris)
plot(model)
}
\seealso{
\code{\link{OneR}}
}

