% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneR.R
\name{print.OneR}
\alias{print.OneR}
\title{Print OneR models}
\usage{
\method{print}{OneR}(x, ...)
}
\arguments{
\item{x}{object of class \code{"OneR"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} method for class \code{OneR}.
}
\details{
Prints the rules and the accuracy of an OneR model.
}
\examples{
model <- OneR(iris)
print(model)
}
\author{
Holger von Jouanne-Diedrich
}
\references{
\url{https://github.com/vonjd/OneR}
}
\seealso{
\code{\link{OneR}}
}

