% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneR.R
\name{summary.OneR}
\alias{summary.OneR}
\title{Summarize OneR models}
\usage{
\method{summary}{OneR}(object, ...)
}
\arguments{
\item{object}{object of class \code{"OneR"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{summary} method for class \code{OneR}.
}
\details{
Prints the rules of the OneR model, the accuracy, a contingency table of the feature attribute and the target and performs a chi-squared test on this table.

In the contingency table the maximum values in each column are highlighted by adding a '*', thereby representing the rules of the OneR model.
}
\examples{
model <- OneR(iris)
summary(model)
}
\author{
Holger von Jouanne-Diedrich, \email{r-project@ephorie.de}
}
\references{
\url{http://vonjd.github.io/OneR/}
}
\seealso{
\code{\link{OneR}}
}
\keyword{diagnostics}

