% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneR.R
\name{is.OneR}
\alias{is.OneR}
\title{Test OneR model objects}
\usage{
is.OneR(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Test if object is a OneR model.
}
\examples{
model <- OneR(iris)
is.OneR(model) # evaluates to TRUE
}
\author{
Holger von Jouanne-Diedrich, \email{r-project@ephorie.de}
}
\keyword{OneR}
\keyword{model}

